Dimensionality_Reduction/requirements.txt
streamlit==1.22.0
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.2.2
plotly==5.13.1
matplotlib==3.7.1
seaborn==0.12.2
scikit-image==0.20.0
pillow==9.5.0 